Fixed typecast errors since move to go 1.18

+++ pkg/grpc/grpc_test.go
@@ -70,7 +70,7 @@ func TestGRPCCommunication(t *testing.T) {
 	assert.EqualValues(t, 456, r.Bytes)
 	assert.EqualValues(t, 0x11223344, r.GetNetwork().GetSrcAddr().GetIpv4())
 	assert.EqualValues(t, 0x55667788, r.GetNetwork().GetDstAddr().GetIpv4())
-	assert.EqualValues(t, 0xaabbccdd, r.GetAgentIp().GetIpv4())
+	assert.EqualValues(t, uint64(0xaabbccdd), r.GetAgentIp().GetIpv4())
 	select {
 	case rs = <-serverOut:
 	case <-time.After(timeout):
@@ -82,8 +82,8 @@ func TestGRPCCommunication(t *testing.T) {
 	assert.EqualValues(t, 1, r.Flags)
 	assert.EqualValues(t, 101, r.Bytes)
 	assert.EqualValues(t, 0x44332211, r.GetNetwork().GetSrcAddr().GetIpv4())
-	assert.EqualValues(t, 0x88776655, r.GetNetwork().GetDstAddr().GetIpv4())
-	assert.EqualValues(t, 0xddccbbaa, r.GetAgentIp().GetIpv4())
+	assert.EqualValues(t, uint64(0x88776655), r.GetNetwork().GetDstAddr().GetIpv4())
+	assert.EqualValues(t, uint64(0xddccbbaa), r.GetAgentIp().GetIpv4())
 
 	select {
 	case rs = <-serverOut:
